Allow calls in other directories
This failed if the call didn't happen in the same directory (or somewhere in that repository) as file we were trying to compute the median age of. This fixes that by `cd`ing into the directory containing the file, and then calling `git blame`.

I put all this in a scope (using parentheses) so that we can `cd` into the new directory and not need to worry about `cd`ing back to the initial directory.