Output the timestamp in a human readable form
This captures the timestamp (using `$()`) and then uses `date -d` to convert that from a Unix timecode to a more human readable form.

I'm not sure we _always_ want this output in this form; there might be times when a time stamp might be more useful (for sorting, e.g.).

It would also be nice to (have the option to) include the file names in the output and handle directories, lists of files, etc.