Extend the script to also get other statistics
This extends the script so we also get `min`, `max`, and `sstdev` (all of which are easy in `datamash`). The downside is that we can't easily do the conversion of the unix time stamps to human readable dates.

We'd probably need to introduce flags at this point to try to make something more useful, and that was more than we really wanted to be doing here, so I think we'll stop now.